Add target 'cuda' to Makefile
The nvcc linker is stupid and won't recognize functions in other files linked
with gcc or other files linked with nvcc that end in .c.  So, .cu versions of
all relevant files for the CUDA version have been copied from the dependencies
in the sequential version. Let's hope we can fix this hack later.

Note that all the CUDA source files have unique prefix names so that the object
names don't overlap with those of the sequential version. This way, we don't
break 'make seq'.