First swing at CUDA --> ONLY WORKS FOR SMALL INPUT
find_nearest_cluster() has been moved to the GPU, but the implementation is not
complete. Currently, the code launches only one block with a thread for each
object. Since CUDA does not support arbitrarily large block sizes, the block
size is fixed to 128 threads.

Everything else is still done on the CPU.